requirements.txt
dwave-ocean-sdk
numpy~=1.25.0